refactor(agent,pkg): rename server package to SSH
