+++ pkg/agent/ssh/modes/host/command/command_native.go
