+++ pkg/agent/ssh/modes/host/sessioner.go
@@ -12,9 +12,9 @@ import (
 
 	gliderssh "github.com/gliderlabs/ssh"
 	"github.com/shellhub-io/shellhub/pkg/agent/pkg/osauth"
-	"github.com/shellhub-io/shellhub/pkg/agent/server/modes"
-	"github.com/shellhub-io/shellhub/pkg/agent/server/modes/host/command"
-	"github.com/shellhub-io/shellhub/pkg/agent/server/utmp"
+	"github.com/shellhub-io/shellhub/pkg/agent/ssh/modes"
+	"github.com/shellhub-io/shellhub/pkg/agent/ssh/modes/host/command"
+	"github.com/shellhub-io/shellhub/pkg/agent/ssh/utmp"
 	log "github.com/sirupsen/logrus"
 	gossh "golang.org/x/crypto/ssh"
 )
