+++ pkg/agent/ssh/modes/mocks/authenticator.go
