+++ pkg/agent/ssh/server.go
@@ -1,4 +1,4 @@
-package server
+package ssh
 
 import (
 	"net"
@@ -7,8 +7,8 @@ import (
 	"time"
 
 	gliderssh "github.com/gliderlabs/ssh"
-	"github.com/shellhub-io/shellhub/pkg/agent/server/modes"
-	"github.com/shellhub-io/shellhub/pkg/agent/server/modes/host"
+	"github.com/shellhub-io/shellhub/pkg/agent/ssh/modes"
+	"github.com/shellhub-io/shellhub/pkg/agent/ssh/modes/host"
 	"github.com/shellhub-io/shellhub/pkg/api/client"
 	log "github.com/sirupsen/logrus"
 	gossh "golang.org/x/crypto/ssh"
