+++ pkg/agent/ssh/sftp.go
@@ -1,4 +1,4 @@
-package agent
+package ssh
 
 import (
 	"errors"
@@ -9,7 +9,7 @@ import (
 	"syscall"
 
 	"github.com/pkg/sftp"
-	"github.com/shellhub-io/shellhub/pkg/agent/server/modes/host/command"
+	"github.com/shellhub-io/shellhub/pkg/agent/ssh/modes/host/command"
 )
 
 type pipe struct {
